refactor: refactor some functions and remove unnessessary code
