+++ package.json
@@ -3,12 +3,12 @@
     "version": "1.0.0",
     "description": "Template for the WebSocket remote control task",
     "main": "index.ts",
-
     "scripts": {
         "start": "ts-node ./index.ts --watch ./src",
         "start:dev": "nodemon",
         "start:prod": "webpack --mode=production && node build/index.cjs",
-        "test": "echo \"Error: no test specified\" && exit 1"
+        "test": "echo \"Error: no test specified\" && exit 1",
+        "lint": "npx eslint ./src"
     },
     "author": "Andrei Auchynnikau",
     "keywords": [],
