+++ src/http_server/index.ts
@@ -15,16 +15,3 @@ export const httpServer = http.createServer(function (req, res) {
 		res.end(data);
 	});
 });
-
-
-/*const wss = new WebSocketServer({ httpServer });
-
-wss.on("connection", function connection(ws) {
-	ws.on("error", console.error);
-
-	ws.on("message", function message(data) {
-		console.log("received: %s", data);
-	});
-
-	ws.send("something");
-});*/
