--- src/websocket_server/examples.ts
@@ -1,7 +1,0 @@
-const addShips = {
-	type: "add_ships",
-	data: "{\"gameId\":1,\"ships\":[{\"position\":{\"x\":2,\"y\":2},\"direction\":false,\"type\":\"small\",\"length\":1},{\"position\":{\"x\":3,\"y\":6},\"direction\":false,\"type\":\"small\",\"length\":1},{\"position\":{\"x\":5,\"y\":6},\"direction\":false,\"type\":\"small\",\"length\":1},{\"position\":{\"x\":7,\"y\":2},\"direction\":false,\"type\":\"large\",\"length\":3},{\"position\":{\"x\":1,\"y\":5},\"direction\":true,\"type\":\"huge\",\"length\":4},{\"position\":{\"x\":8,\"y\":6},\"direction\":true,\"type\":\"large\",\"length\":3},{\"position\":{\"x\":4,\"y\":1},\"direction\":true,\"type\":\"medium\",\"length\":2},{\"position\":{\"x\":8,\"y\":4},\"direction\":false,\"type\":\"medium\",\"length\":2},{\"position\":{\"x\":3,\"y\":4},\"direction\":false,\"type\":\"medium\",\"length\":2},{\"position\":{\"x\":0,\"y\":0},\"direction\":false,\"type\":\"small\",\"length\":1}],\"indexPlayer\":1}",
-	id: 0
-};
-
-const a = { type: 123, data:  [{id: 1, name: "adam"}, {id: 2, name: "smit"}]};
